Extract user mod select button for reuse
